backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-genai>=1.0.0
google-generativeai>=0.8.0
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
